+++ packages/client/src/hooks/useValidation.ts
@@ -1,4 +1,4 @@
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 
 export type ValidationData = {
   field: string;
@@ -6,71 +6,29 @@ export type ValidationData = {
   text?: string;
 };
 
-export type ValidationRule = {
-  field: string;
-  validation: (value?: string) => { isValid: boolean; text: string };
-};
-
-export function useValidation(validations: ValidationRule[]) {
-  const [validationData, setValidationData] = useState<
-    Record<string, ValidationData>
-  >({});
-
-  const isFormValid = useMemo(() => {
-    return !Object.entries(validationData).some(d => !d[1].isValid);
-  }, [validationData]);
+export function useValidation(
+  validations: { field: string; validation: any }[]
+) {
+  const [validationData, setValidationData] = useState<ValidationData[]>([]);
 
   function validateForm(values: Record<string, string | undefined>) {
-    const newValidationData: Record<string, ValidationData> = {};
-    let isValid = true;
+    const newValidationData: ValidationData[] = [];
 
     for (const validation of validations) {
-      newValidationData[validation.field] = {
+      newValidationData.push({
         ...validation.validation(values[validation.field]),
         field: validation.field,
-      };
-
-      if (isValid && !newValidationData[validation.field].isValid) {
-        isValid = false;
-      }
+      });
     }
 
     setValidationData(newValidationData);
 
-    return isValid;
-  }
-
-  function validateField(field: string, value?: string) {
-    const newValidationData: Record<string, ValidationData> = {
-      ...validationData,
-    };
-    delete newValidationData[field];
-    const validation = validations.find(v => v.field === field);
-
-    if (validation) {
-      newValidationData[validation.field] = {
-        ...validation.validation(value),
-        field: validation.field,
-      };
-    }
-
-    setValidationData(newValidationData);
+    return !newValidationData.some(v => v.isValid === false);
   }
 
   function clearFieldValidation(field: string) {
-    const newValidationData: Record<string, ValidationData> = {
-      ...validationData,
-    };
-    delete newValidationData[field];
-
-    setValidationData(newValidationData);
+    setValidationData(validationData.filter(d => d.field !== field));
   }
 
-  return {
-    validationData,
-    isFormValid,
-    validateForm,
-    validateField,
-    clearFieldValidation,
-  };
+  return { validationData, validateForm, clearFieldValidation };
 }
